feat: the 32-bit lane — u32 arithmetic + u32-in-state (ABI v2, .cell v4)
Compiler (rustz80). Full u32 arithmetic: add/sub as inline ADD/ADC / SBC carry
chains (authentic on both targets); mul/div/rem via new ED FE traps 0x12/0x13 on
the Cell target and real software siblings on Spectrum — __mul32 by three 16-bit
partial products over a new __mul16w 16x16->32 core, __divmod32 restoring
division with a 33rd-bit forced-commit path (exact even for divisors >= 2^31),
both emitted through the Asm with labels rather than hand-counted offsets.
Mixed 16-bit operands zero-extend (the unsuffixed-literal mixing rustc allows);
wrapping_* maps onto the mod-2^32 ops. u32 struct fields: two little-endian
slots (FieldDef.width distinguishes them from 2-element arrays), wide access by
value (Var32/Assign32) and through the self pointer (new Deref32/Store32).
Every u32-into-16-bit context (returns, params, call args, comparisons,
conditions, stores, for-bounds) is now a clean lowering error, never a codegen
panic — also fixing two latent miscompiles (u16var = u32expr corrupted the
neighbouring slot; u32var = 16-bit expr panicked) and making `let x: u32 = 5`
respect its annotation.

Cell layer (cell80). The typed-state ABI carries a width per field:
state_addrs is (name, addr, Ty), .cell format v4 (back-compat reads of v3/v2);
StateCell and CellHost::run_state drive and read u32 fields by name (PyO3/MCP
were already u64 end-to-end). ABI_VERSION 2: the traps are additive on a v2
host, but a v1 host no-ops unknown ids, so a wide cartridge must declare what it
needs (docs/09). CellGraph routes u32 edges — u32->u32 wires type-check,
u32->u16 narrowing is rejected before a single cycle runs.

Library (100 cells). percent/permille/ratio_255/scale_percent/within_percent
compute wide and saturate (their intermediate overflows are gone); euclid_sq's
dist is a wide u32 field; new wide siblings square_wide / weighted_sum_wide.
The overflow_audit example now shows the whole arc: 5/7 value cells fixed by
u32 arithmetic, the 2 result-overflow cells honestly capped by their u16
return, and 3/3 wide state fields exact — compute wide, persist wide, read
wide by name.

+++ cell80/tests/cell_fuzz.rs
@@ -80,25 +80,37 @@ fn state_named_roundtrip_fuzz() {
     for _ in 0..500 {
         let a = (rng.next() & 0xFFFF) as u16;
         let b = (rng.next() & 0xFFFF) as u16;
-        cell.set("a", a).unwrap(); // queued; applied to memory by run()
-        cell.set("b", b).unwrap();
+        cell.set("a", a as u64).unwrap(); // queued; applied to memory by run()
+        cell.set("b", b as u64).unwrap();
         cell.run(DEFAULT_CYCLES).unwrap();
         // input half: the queued inputs landed in memory and read back by name unchanged
         assert_eq!(
             cell.get("a"),
-            Some(a),
+            Some(a as u64),
             "input `a` did not round-trip by name"
         );
         assert_eq!(
             cell.get("b"),
-            Some(b),
+            Some(b as u64),
             "input `b` did not round-trip by name"
         );
         // output half: every output (incl. one written last) read back by name == host oracle
-        assert_eq!(cell.get("sum"), Some(a.wrapping_add(b)), "sum ({a},{b})");
-        assert_eq!(cell.get("diff"), Some(a.wrapping_sub(b)), "diff ({a},{b})");
-        assert_eq!(cell.get("prod"), Some(a.wrapping_mul(b)), "prod ({a},{b})");
-        assert_eq!(cell.get("big"), Some(a.max(b)), "big ({a},{b})");
+        assert_eq!(
+            cell.get("sum"),
+            Some(a.wrapping_add(b) as u64),
+            "sum ({a},{b})"
+        );
+        assert_eq!(
+            cell.get("diff"),
+            Some(a.wrapping_sub(b) as u64),
+            "diff ({a},{b})"
+        );
+        assert_eq!(
+            cell.get("prod"),
+            Some(a.wrapping_mul(b) as u64),
+            "prod ({a},{b})"
+        );
+        assert_eq!(cell.get("big"), Some(a.max(b) as u64), "big ({a},{b})");
     }
 }
